@@ -1,22 +1,22 @@
 name: invalid
 description: A new Flutter package project.
-version: 0.0.1
+version: 1.0.0
 author: null
 homepage: null
 
 environment:
   sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.17.0 <2.0.0"
+  flutter: ">=2.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  bloc: ^6.0.3
-  equatable: ^1.2.5
-  intl: ^0.16.1
-  flutter_bloc: ^6.0.5
+  bloc: ^6.1.3
+  equatable: ^2.0.0
+  intl: ^0.17.0
+  flutter_bloc: ^6.1.3
 
 dev_dependencies:
-  bloc_test: ^7.0.4
+  bloc_test: ^7.1.0
   flutter_test:
     sdk: flutter
